refactor(sidecar): the actions move to the products that answer them
terra/actions.py is gone. Each of the seventeen actions now sits in the slice
that answers it, the registry names it there, and adding a product is one
directory and one line:

  terra/landcover/actions.py  predict, lulc, domain_shift, domain_shift_cohort
  terra/energy/actions.py     solar_resource, solar_terrain, solar_siting,
                              energy_model, wind_resource
  terra/canopy/actions.py     canopy_field, canopy_mesh, canopy_from_aoi
  terra/water/actions.py      water
  terra/flood/actions.py      flood_envelope
  terra/scenes/actions.py     list_datacube, render_composite
  terra/cli.py                ping

terra/scenes/ is new and holds only actions: both answer about the imagery
itself rather than about anything derived from it. They sit there rather than
inside terra.imagery because that package is a reader, and a reader with
actions in it stops being one. ping joins the entry point, because it answers
about the process.

WHAT THE CONTRACT REFUSED, AND WHAT IT WAS RIGHT ABOUT. With the actions inside
the slices there is no longer a neutral place for an edge between two products,
and three appeared at once: energy reaching land cover for a MapBiomas window,
land cover reaching canopy for the crop class set, and land cover reaching
canopy for phenology.

None of the three was a product depending on a product. MapBiomas is an
external source, and it was split across three packages -- the window fetched
from the Brazil COG in one, the legend and colours in another, which ids count
as cropland in a third. Every one of those is a fact about the source. They are
now terra/mapbiomas.py, one reader for one source, beside terra/stac.py.

Phenology reads a vegetation cycle out of an index series. The classification
attaches it to its payload and the canopy uses it to size a season; it is below
both, as terra/phenology.py.

terra.aoi and terra.phenology join the layers list. They were unlisted, and an
unlisted module is one the contract does not constrain -- which is how the
edges that broke it could have been written there instead.

Verification: pytest 531 passed, 8 skipped. ruff check clean, three contracts
kept, gofmt, go vet and go test -race clean across every package. A ping
through the entry point answers, and every action resolves to the slice that
owns it.
