+++ sidecar/terra/landcover/spectra.py
@@ -17,7 +17,7 @@ from pathlib import Path
 import numpy as np
 
 from terra.imagery import sentinel2
-from terra.landcover.palette import (
+from terra.mapbiomas import (
     CLASSIFIER_COLORS as MAPBIOMAS_COLORS,
     CLASSIFIER_LEGEND as MAPBIOMAS_LEGEND,
 )
