+++ sidecar/terra/scenes/__init__.py
@@ -1,0 +1,9 @@
+"""
+The scenes themselves: which acquisitions cover an area, and what one looks
+like.
+
+The only slice with no module of its own beyond its actions, because both
+answers are the reader's output arranged for the caller rather than a
+computation over it. It sits here rather than inside terra.imagery so that
+package stays a reader.
+"""
